requirements.txt
pygame>=2.5.0
PyYAML>=6.0.1